refactor(donate): separeted by files
