+++ src/ai_telegram_bot/handlers/user/__init__.py
@@ -1,13 +1,13 @@
 from aiogram import F, Router
 from aiogram.filters import Command, CommandStart, StateFilter
+from donate import prepare_router as prepare_donate_router
 
 from ai_telegram_bot import states
 from ai_telegram_bot.filters import ChatTypeFilter, TextFilter
 
-from . import clear, donate, start, taro, text, voice
+from . import clear, start, taro, text, voice
 
 
-# TODO: extract donate to folder
 def prepare_router() -> Router:
     user_router = Router(name="user_router")
     user_router.message.filter(ChatTypeFilter("private"))
@@ -24,13 +24,6 @@ def prepare_router() -> Router:
     user_router.message.register(text.handle_message, F.text)
     user_router.message.register(voice.handle_voice_message, F.voice)
 
-    user_router.message.register(donate.cmd_donate, Command("donate"))
-    user_router.message.register(donate.cmd_paysupport, Command("paysupport"))
-    user_router.message.register(donate.cmd_refund, Command("refund"))
-    user_router.pre_checkout_query.register(donate.pre_checkout_query)
-    user_router.callback_query.register(
-        donate.on_donate_cancel, F.data == "donate_cancel"
-    )
-    user_router.message.register(donate.on_successfull_payment, F.successful_payment)
+    user_router.include_router(prepare_donate_router())
 
     return user_router
