+++ src/ai_telegram_bot/handlers/user/donate/paysupport.py
@@ -1,0 +1,6 @@
+from aiogram.types import Message
+from fluent.runtime import FluentLocalization
+
+
+async def cmd_paysupport(message: Message, l10n: FluentLocalization) -> None:
+    await message.answer(l10n.format_value("donate-paysupport-message"))
